split fungible asset v2 to a different processor
