+++ rust/processor/src/processors/mod.rs
@@ -3,18 +3,21 @@
 
 pub mod coin_processor;
 pub mod default_processor;
+pub mod fungible_asset_processor;
 pub mod processor_trait;
 pub mod stake_processor;
 pub mod token_processor;
 
 use self::{
     coin_processor::NAME as COIN_PROCESSOR_NAME, default_processor::NAME as DEFAULT_PROCESSOR_NAME,
+    fungible_asset_processor::NAME as FUNGIBLE_ASSET_PROCESSOR_NAME,
     stake_processor::NAME as STAKE_PROCESSOR_NAME, token_processor::NAME as TOKEN_PROCESSOR_NAME,
 };
 
 pub enum Processor {
     CoinProcessor,
     DefaultProcessor,
+    FungibleAssetProcessor,
     StakeProcessor,
     TokenProcessor,
 }
@@ -24,6 +27,7 @@ impl Processor {
         match input_str.as_str() {
             DEFAULT_PROCESSOR_NAME => Self::DefaultProcessor,
             COIN_PROCESSOR_NAME => Self::CoinProcessor,
+            FUNGIBLE_ASSET_PROCESSOR_NAME => Self::FungibleAssetProcessor,
             STAKE_PROCESSOR_NAME => Self::StakeProcessor,
             TOKEN_PROCESSOR_NAME => Self::TokenProcessor,
             _ => panic!("Processor unsupported {}", input_str),
